Add: Spack
Spack is a flexible package manager that supports multiple versions,
configurations, platforms, and compilers. It's perfectly suited for
complex software toolchains as often seen on clusters and seamlessly
allows to combine C/C++, Fortran, Python, R, ... projects and
bindings with a high user-control on variants, flags and dependencies
of all packages.