Update entry for morphologica (now mathplot) (#1705)
* Update entry for morphologica (now mathplot)

In July, morphologica was renamed to mathplot to better describe the
reason for the library's existence.

I quick look at

https://github.com/ABRG-Models/morphologica/

will confirm that morphologica has been archived and development moved to

https://github.com/sebsjames/mathplot

* Adds entry for sebsjames/maths

I'm submitting the maths library that is used in mathplot. The library
is mature and continuously developed. It began life as part of
[morphologica](https://github.com/ABRG-Models/morphologica) and was
split out into its own repository when morphologica was renamed to
[mathplot](https://github.com/sebsjames/mathplot)